Rearranged the file so I KNOW THE CODE WORKS
