+++ main.py
@@ -1,0 +1,4 @@
+from gui import start_gui
+
+if __name__ == "__main__":
+    start_gui()
