--- src/road_to_waffle_house/main.py
@@ -1,4 +1,0 @@
-from src.road_to_waffle_house.gui import start_gui
-
-if __name__ == "__main__":
-    start_gui()
